# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Routeyou events API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/events/{event}/achievements/{qrcode}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Complete achievement for a checkpoint",
                "operationId": "91024944c2c2eb5123d163140d2f657c",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "qrcode",
                        "in": "path",
                        "description": "QR code of the checkpoint",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Achievement completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The achievement has been updated successfully."
                                        },
                                        "achievement": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "completed": true,
                                                "checkpoint_id": 1,
                                                "participation_id": 1
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not allowed to access this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The checkpoint or participation was not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": [
                            "USER"
                        ]
                    }
                ]
            }
        },
...
        "/api/routes/{route}/checkpoints": {
            "post": {
                "tags": [
                    "Routes"
                ],
                "summary": "Create a checkpoint and add it to a route",
                "operationId": "234eef5ed41abe4660e82d91322dc1fd",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "ID of the route to which the checkpoint will be added",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "longitude": {
                                        "type": "number",
                                        "format": "double",
                                        "maximum": "180",
                                        "minimum": "-180",
                                        "example": "45.1234"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "double",
                                        "maximum": "90",
                                        "minimum": "-90",
                                        "example": "-75.6789"
                                    },
                                    "coin": {
                                        "type": "number",
                                        "maximum": "1000",
                                        "example": "500"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Checkpoint created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Checkpoint created"
                                        },
                                        "checkpoint": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "longitude": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "example": "45.1234"
                                                },
                                                "latitude": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "example": "-75.6789"
                                                },
                                                "coin": {
                                                    "type": "number",
                                                    "example": "500"
                                                },
                                                "qr_code": {
                                                    "type": "string",
                                                    "example": "app_url/storage/qr-codes/9dca6322-1f2d-4777-a352-7310fcd2ff10.svg"
                                                },
                                                "route_id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not allowed to access this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Route not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Route not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "longitude": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The longitude field is required."
                                                    }
                                                },
                                                "latitude": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The latitude field is required."
                                                    }
                                                },
                                                "coin": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The coin field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": [
                            "ORGANISER"
                        ]
                    }
                ]
            }
        },
        "/api/routes/{route}/checkpoints/{checkpoint}": {
            "delete": {
                "tags": [
                    "Routes"
                ],
                "summary": "Delete a checkpoint from a route",
                "operationId": "f0bbd38607a6f7acda226a2fcc4c7cfd",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "ID of the route from which the checkpoint will be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "checkpoint",
                        "in": "path",
                        "description": "ID of the checkpoint to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checkpoint deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Checkpoint deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not allowed to access this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Checkpoint or route not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Checkpoint or route not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": [
                            "ORGANISER"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}